Update documentation and minor clean up in triqs/utility
- add doxygen documentation to all files
- fix includes
- remove clangd and clang-tidy warnings
- some code modernization

+++ c++/triqs/utility/chebyshev.hpp
@@ -30,7 +30,12 @@
 namespace triqs::utility {
 
   /**
-   * @brief Compute Chebyshev points of the first kind on the interval [-1, 1].
+   * @ingroup triqs-utility-math
+   * @{
+   */
+
+  /**
+   * @brief Compute Chebyshev points of the first kind on the interval \f$ [-1, 1] \f$.
    *
    * @details Points are computed as:
    * \f[
@@ -40,14 +45,14 @@ namespace triqs::utility {
    * Points are returned in ascending order.
    *
    * @param N Number of Chebyshev points.
-   * @return Vector of N Chebyshev points in ascending order on [-1, 1].
+   * @return Vector of \f$ N \f$ Chebyshev points in ascending order on \f$ [-1, 1] \f$.
    */
   inline nda::vector<double> chebyshev_points(long N) {
     nda::vector<double> points(N);
     const double pi = std::numbers::pi;
     for (long i = 0; i < N; ++i) {
       // Store in ascending order (cos decreases as argument increases)
-      points[N - 1 - i] = std::cos((2.0 * i + 1.0) * pi / (2.0 * N));
+      points[N - 1 - i] = std::cos((2.0 * static_cast<double>(i) + 1.0) * pi / (2.0 * static_cast<double>(N)));
     }
     return points;
   }
@@ -63,37 +68,39 @@ namespace triqs::utility {
    * Weights are returned in the same order as chebyshev_points().
    *
    * @param N Number of Chebyshev points.
-   * @return Vector of N barycentric weights.
+   * @return Vector of \f$ N \f$  barycentric weights.
    */
   inline nda::vector<double> chebyshev_barycentric_weights(long N) {
     nda::vector<double> weights(N);
     const double pi = std::numbers::pi;
     for (long i = 0; i < N; ++i) {
       // Match point ordering (ascending)
-      double sign = (i % 2 == 0) ? 1.0 : -1.0;
-      weights[N - 1 - i] = sign * std::sin((2.0 * i + 1.0) * pi / (2.0 * N));
+      double sign        = (i % 2 == 0) ? 1.0 : -1.0;
+      weights[N - 1 - i] = sign * std::sin((2.0 * static_cast<double>(i) + 1.0) * pi / (2.0 * static_cast<double>(N)));
     }
     return weights;
   }
 
   /**
-   * @brief Scale a value from interval [a, b] to the standard interval [-1, 1].
+   * @brief Scale a value from interval \f$ [a, b] \f$ to the standard interval \f$ [-1, 1] \f$.
    *
-   * @param x Value in [a, b].
+   * @param x Value in \f$ [a, b] \f$.
    * @param a Lower bound of the original interval.
    * @param b Upper bound of the original interval.
-   * @return Scaled value in [-1, 1].
+   * @return Scaled value in \f$ [-1, 1] \f$.
    */
   inline double to_standard_interval(double x, double a, double b) { return 2.0 * (x - 0.5 * (b + a)) / (b - a); }
 
   /**
-   * @brief Scale a value from the standard interval [-1, 1] to interval [a, b].
+   * @brief Scale a value from the standard interval \f$ [-1, 1] \f$ to interval \f$ [a, b] \f$.
    *
-   * @param x Value in [-1, 1].
+   * @param x Value in \f$ [-1, 1] \f$.
    * @param a Lower bound of the target interval.
    * @param b Upper bound of the target interval.
-   * @return Scaled value in [a, b].
+   * @return Scaled value in \f$ [a, b] \f$.
    */
   inline double from_standard_interval(double x, double a, double b) { return 0.5 * (b - a) * x + 0.5 * (b + a); }
+
+  /** @} */
 
 } // namespace triqs::utility
